+++ src/Andromeda/AvaloniaApp.FSharp/Helpers/NotificationData.fs
@@ -1,0 +1,4 @@
+namespace Andromeda.AvaloniaApp.FSharp.Helpers
+
+type NotificationData (message: string) =
+    member val Message = message with get,set