+++ src/Andromeda/Core/Constants.fs
@@ -1,0 +1,6 @@
+namespace Andromeda.Core.FSharp
+
+[<RequireQualifiedAccess>]
+module Constants =
+    let folderName = "andromeda"
+    let databaseFile = "andromeda.db"