+++ src/Andromeda/Core/Database.fs
@@ -1,0 +1,10 @@
+namespace Andromeda.Core.FSharp
+
+open System.IO
+
+open Andromeda.Core.FSharp
+
+module Database =
+    Directory.CreateDirectory(SystemInfo.savePath) |> ignore
+
+    let name = Path.Combine(SystemInfo.savePath, Constants.databaseFile)