+++ src/Andromeda/Core/Path.fs
@@ -1,0 +1,1 @@
+module Andromeda.Core.FSharp.Path