+++ src/Andromeda/Core/PersistenceTypes.fs
@@ -1,0 +1,11 @@
+namespace Andromeda.Core.FSharp
+
+module PersistenceTypes =
+    // TODO: Add Uniontypes for errors and return Result types
+    type LoadAuthorizationData = unit -> Authentication
+
+    type SaveAuthorizationData = Authentication -> bool
+
+    type LoadSettings = unit -> Settings
+
+    type SaveSettings = Settings -> unit