+++ src/Andromeda/Core/User.fs
@@ -1,0 +1,3 @@
+module Andromeda.Core.FSharp.User
+
+let getUserData = GogApi.DotNet.FSharp.User.askForUserData